backend/requirements.txt
Django>=4.2.2
psycopg2>=2.9.6
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
dj-rest-auth>=4.0.1
dj-rest-auth[with_social]
django-allauth>=0.54.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.1.0
pgvector
pre-commit
moviepy
openai
tenacity
pandas
factory_boy
faker
black
django-storages[boto3]
